Fix DictRefsContainer.set_if_equals() to only update requested ref
The set_if_equals() method was incorrectly updating all refs in a
symbolic reference chain instead of just the requested ref. This was
because it used follow() to get all refs in the chain and then updated
all of them.

This caused issues where updating a symbolic ref would incorrectly
modify the ref it pointed to.